requirements-render.txt
anthropic==0.19.1
beautifulsoup4==4.12.2
email-validator==2.1.1
flask==3.0.2
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
google-generativeai==0.3.2
gunicorn==21.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
psycopg2-binary==2.9.9
scikit-learn==1.4.1.post1
scipy==1.12.0
sqlalchemy==2.0.28
werkzeug==3.0.1
wtforms==3.1.2
